// src/app/core/api/model/appLogEntryDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AppLogEntryDto { 
    timestamp: string;
    level?: string;
    /**
     * Log message
     */
    message: string;
    namespace?: string;
    app?: string;
    pod?: string;
    container?: string;
    stream?: string;
    server_id?: string;
    hostname?: string;
    server_type?: string;
    /**
     * Raw parsed JSON metadata (if log line is JSON)
     */
    metadata?: object;
}