// src/app/core/api/model/appLogVolumeResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LogVolumeLevelSeriesDto } from './logVolumeLevelSeriesDto';


export interface AppLogVolumeResponseDto { 
    cluster_id: string;
    namespace?: string;
    app?: string;
    range_start: string;
    range_end: string;
    /**
     * Bucket step size
     */
    step: string;
    /**
     * One time series per log level
     */
    series: Array<LogVolumeLevelSeriesDto>;
    queried_at: string;
}